requirements.txt
line-bot-sdk
flask
requests
colorthief
pyrebase